@@ -1,7 +1,13 @@
 import type { TuftPreHandler, TuftResponder } from '../src/route-map';
 import type { HttpError } from '../src/utils';
-import { constants } from 'http2';
-import { promises as fsPromises } from 'fs';
+
+import {
+  HTTP_STATUS_OK,
+  HTTP_STATUS_TEAPOT,
+  HTTP_HEADER_CONTENT_TYPE,
+  HTTP_HEADER_ACCEPT_RANGES,
+  HTTP_HEADER_LAST_MODIFIED
+} from '../src/constants';
 import {
   createResponseHandler,
   returnResponse,
@@ -16,78 +22,53 @@ import {
   handleTextResponse,
   handleHtmlResponse,
   handleJsonResponse,
-  statCheck,
-  onError
 } from '../src/response-handlers';
-import {
-  HTTP2_HEADER_STATUS,
-  HTTP2_HEADER_LOCATION,
-  HTTP2_HEADER_CONTENT_TYPE,
-  HTTP2_HEADER_CONTENT_LENGTH,
-  HTTP2_HEADER_METHOD,
-  HTTP2_HEADER_PATH,
-  HTTP2_HEADER_ACCEPT_RANGES,
-  HTTP2_HEADER_LAST_MODIFIED,
-} from '../src/constants';
 
-const {
-  HTTP_STATUS_OK,
-  HTTP_STATUS_FOUND,
-  HTTP_STATUS_BAD_REQUEST,
-  HTTP_STATUS_TEAPOT,
-} = constants;
+const mockConsoleError = jest
+  .spyOn(console, 'error')
+  .mockImplementation(() => { });
 
-const CONTENT_LENGTH = 42;
+const mockExit = jest
+  .spyOn(process, 'exit')
+  .mockImplementation(() => undefined as never);
 
-const mockIncomingHeaders = {
-  [HTTP2_HEADER_METHOD]: 'GET',
-  [HTTP2_HEADER_PATH]: '/',
-};
-
-const mockFileHandle = {
-  stat: async () => {
-    return { size: CONTENT_LENGTH };
-  },
-} as fsPromises.FileHandle;
-
-const mockFsOpen = jest.spyOn(fsPromises, 'open').mockImplementation(async () => mockFileHandle);
-const mockConsoleError = jest.spyOn(console, 'error').mockImplementation(() => { });
-const mockExit = jest.spyOn(process, 'exit').mockImplementation(() => undefined as never);
-
-const mockStream = {
-  respond: jest.fn(),
-  respondWithFile: jest.fn(),
-  end: jest.fn(),
-  emit: jest.fn(),
-};
-
-const mockContext: any = {
-  request: {},
-  outgoingHeaders: {},
-  setHeader: jest.fn((key, value) => {
-    mockContext.outgoingHeaders[key] = value;
-  }),
-};
-
-const mockDate = new Date();
-
-const mockStat = {
-  mtime: {
-    toUTCString: jest.fn(() => {
-      return mockDate.toUTCString();
+function createMockRequest(method: string = 'GET', url: string = '/') {
+  const mockRequest: any = {
+    method,
+    url,
+    on: jest.fn((_, callback) => {
+      callback();
     }),
-  },
-};
+  };
 
-beforeEach(() => {
-  mockStream.respond.mockClear();
-  mockStream.respondWithFile.mockClear();
-  mockStream.end.mockClear();
-  mockStream.emit.mockClear();
-});
+  return mockRequest;
+}
+
+function createMockResponse() {
+  const mockResponse: any = {
+    _headers: {},
+    statusCode: HTTP_STATUS_OK,
+    on: jest.fn(),
+    once: jest.fn(),
+    emit: jest.fn(),
+    write: jest.fn(),
+    end: jest.fn(),
+    setHeader: jest.fn((name: string, value: string | number) => {
+      mockResponse._headers[name] = value;
+      return mockResponse;
+    }),
+    getHeader: jest.fn((name: string) => {
+      return mockResponse._headers[name];
+    }),
+    hasHeader: jest.fn((name: string) => {
+      return mockResponse._headers[name] !== undefined;
+    }),
+  };
+
+  return mockResponse;
+}
 
 afterAll(() => {
-  mockFsOpen.mockRestore();
   mockConsoleError.mockRestore();
   mockExit.mockRestore();
 });
@@ -162,18 +143,18 @@ describe('handleResponseObject()', () => {
     test('resolves to be undefined', async () => {
       const response = {};
       const responders = [jest.fn((response: {}) => response)];
-      const outgoingHeaders = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
-      const result = handleResponseObject(
+      const returnValue = await handleResponseObject(
         response,
         responders,
-        outgoingHeaders,
-        //@ts-expect-error
-        mockStream,
+        mockRequest,
+        mockResponse,
       );
 
-      await expect(result).resolves.toBeUndefined();
-      expect(responders[0]).toHaveBeenCalledWith(response, mockStream, outgoingHeaders);
+      expect(returnValue).toBeUndefined();
+      expect(responders[0]).toHaveBeenCalledWith(response, mockResponse);
       expect(responders[0]).toHaveReturnedWith(response);
     });
   });
@@ -182,18 +163,18 @@ describe('handleResponseObject()', () => {
     test('resolves to be undefined', async () => {
       const response = {};
       const responders = [jest.fn(() => { })];
-      const outgoingHeaders = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
-      const result = handleResponseObject(
+      const returnValue = await handleResponseObject(
         response,
         responders,
-        outgoingHeaders,
-        //@ts-expect-error
-        mockStream,
+        mockRequest,
+        mockResponse,
       );
 
-      await expect(result).resolves.toBeUndefined();
-      expect(responders[0]).toHaveBeenCalledWith(response, mockStream, outgoingHeaders);
+      expect(returnValue).toBeUndefined();
+      expect(responders[0]).toHaveBeenCalledWith(response, mockResponse);
       expect(responders[0]).toHaveReturned();
       expect(responders[0]).not.toHaveReturnedWith(response);
     });
@@ -213,15 +194,16 @@ describe('handleResponseHandler()', () => {
       const preHandlers: TuftPreHandler[] = [];
       const responders: TuftResponder[] = [];
       const contextOptions = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = handleResponseHandler(
         handler,
         preHandlers,
         responders,
         contextOptions,
-        //@ts-expect-error
-        mockStream,
-        mockIncomingHeaders,
+        mockRequest,
+        mockResponse
       );
 
       await expect(result).resolves.toBeUndefined();
@@ -238,6 +220,8 @@ describe('handleResponseHandler()', () => {
       const preHandlers: TuftPreHandler[] = [];
       const responders: TuftResponder[] = [];
       const contextOptions = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = handleResponseHandler(
         //@ts-ignore
@@ -245,9 +229,8 @@ describe('handleResponseHandler()', () => {
         preHandlers,
         responders,
         contextOptions,
-        //@ts-ignore
-        mockStream,
-        mockIncomingHeaders,
+        mockRequest,
+        mockResponse,
       );
 
       await expect(result).rejects.toThrow('\'null\' is not a valid Tuft response object.');
@@ -269,15 +252,16 @@ describe('handleResponseHandler()', () => {
       ];
       const responders: TuftResponder[] = [];
       const contextOptions = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = handleResponseHandler(
         handler,
         preHandlers,
         responders,
         contextOptions,
-        //@ts-expect-error
-        mockStream,
-        mockIncomingHeaders,
+        mockRequest,
+        mockResponse,
       );
 
       await expect(result).resolves.toBeUndefined();
@@ -294,15 +278,16 @@ describe('handleResponseHandler()', () => {
       const preHandlers = [jest.fn()];
       const responders: TuftResponder[] = [];
       const contextOptions = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = handleResponseHandler(
         handler,
         preHandlers,
         responders,
         contextOptions,
-        //@ts-expect-error
-        mockStream,
-        mockIncomingHeaders,
+        mockRequest,
+        mockResponse,
       );
 
       await expect(result).resolves.toBeUndefined();
@@ -320,21 +305,20 @@ describe('handleResponseHandler()', () => {
       const preHandlers: TuftPreHandler[] = [];
       const responders: TuftResponder[] = [jest.fn((response: {}) => response)];
       const contextOptions = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = handleResponseHandler(
         handler,
         preHandlers,
         responders,
         contextOptions,
-        //@ts-expect-error
-        mockStream,
-        mockIncomingHeaders,
+        mockRequest,
+        mockResponse,
       );
 
-      const expectedHeaders = { [HTTP2_HEADER_STATUS]: HTTP_STATUS_OK };
-
       await expect(result).resolves.toBeUndefined();
-      expect(responders[0]).toHaveBeenCalledWith(response, mockStream, expectedHeaders);
+      expect(responders[0]).toHaveBeenCalledWith(response, mockResponse);
       expect(responders[0]).toHaveReturnedWith(response);
     });
   });
@@ -348,21 +332,22 @@ describe('handleResponseHandler()', () => {
       const preHandlers: TuftPreHandler[] = [];
       const responders: TuftResponder[] = [jest.fn(() => { })];
       const contextOptions = {};
+      const mockRequest = createMockRequest();
+      const mockResponse = createMockResponse();
 
       const result = handleResponseHandler(
         handler,
         preHandlers,
         responders,
         contextOptions,
-        //@ts-expect-error
-        mockStream,
-        mockIncomingHeaders,
+        mockRequest,
+        mockResponse,
       );
 
       await expect(result).resolves.toBeUndefined();
-      expect(responders[0]).toHaveBeenCalledWith({}, mockStream, {});
+      expect(responders[0]).toHaveBeenCalledWith(response, mockResponse);
       expect(responders[0]).toHaveReturned();
-      expect(responders[0]).not.toHaveReturnedWith({});
+      expect(responders[0]).not.toHaveReturnedWith(response);
     });
   });
 });
@@ -373,17 +358,15 @@ describe('handleResponseHandler()', () => {
 
 describe('handleUnknownResponse()', () => {
   describe('when passed a response with an `error` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         error: 'TEAPOT' as HttpError,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -391,17 +374,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with a `redirect` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         redirect: '/foo',
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -409,17 +390,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with a `raw` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         raw: Buffer.from('abc'),
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -427,17 +406,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with a `text` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         text: 'abc',
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -445,17 +422,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with an `html` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         html: '<h1>abc</h1>',
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -463,17 +438,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with a `json` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         json: JSON.stringify('abc'),
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -481,17 +454,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with a `file` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         file: __filename,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -499,17 +470,15 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed a response with a `status` property', () => {
-    test('returns undefined', () => {
+    test('returns undefined', async () => {
       const response = {
         status: HTTP_STATUS_TEAPOT,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
@@ -517,23 +486,17 @@ describe('handleUnknownResponse()', () => {
   });
 
   describe('when passed an empty response', () => {
-    test('stream.respond() is called with the expected arguments', () => {
+    test('response.end() is called', async () => {
       const response = {};
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleUnknownResponse(
+      const result = await handleUnknownResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_OK,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toBeCalledWith(expectedHeaders, { endStream: true });
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -544,48 +507,36 @@ describe('handleUnknownResponse()', () => {
 
 describe('handleHttpErrorResponse', () => {
   describe('when passed a valid value for `error`', () => {
-    test('stream.respond() is called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const response = {
         error: 'TEAPOT' as HttpError,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleHttpErrorResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_TEAPOT,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders, { endStream: true });
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('when passed an invalid value for `error`', () => {
-    test('stream.respond() is called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const response = {
         error: 'FOO' as HttpError,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleHttpErrorResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_BAD_REQUEST,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders, { endStream: true });
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -596,26 +547,19 @@ describe('handleHttpErrorResponse', () => {
 
 describe('handleRedirectResponse()', () => {
   describe('when passed a string for the `url` parameter', () => {
-    test('stream.respond() is called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const response = {
         redirect: '/foo',
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleRedirectResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_FOUND,
-        [HTTP2_HEADER_LOCATION]: response.redirect,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders, { endStream: true });
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -626,56 +570,39 @@ describe('handleRedirectResponse()', () => {
 
 describe('handleBufferResponse()', () => {
   describe('with a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const raw = Buffer.from('abc');
       const response = {
         status: HTTP_STATUS_OK,
         raw,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleBufferResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_OK,
-        [HTTP2_HEADER_CONTENT_TYPE]: 'application/octet-stream',
-        [HTTP2_HEADER_CONTENT_LENGTH]: raw.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(raw);
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('without a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const raw = Buffer.from('abc');
       const response = {
         raw,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleBufferResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'application/octet-stream',
-        [HTTP2_HEADER_CONTENT_LENGTH]: raw.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(raw);
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -686,83 +613,57 @@ describe('handleBufferResponse()', () => {
 
 describe('handleTextResponse()', () => {
   describe('with a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const text = 'abc';
       const response = {
         status: HTTP_STATUS_OK,
         text,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleTextResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_OK,
-        [HTTP2_HEADER_CONTENT_TYPE]: 'text/plain; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: text.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(text);
-
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('without a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const text = 'abc';
       const response = {
         text,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleTextResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'text/plain; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: text.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(text);
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('when passed a boolean value', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const text = true;
       const response = {
         text,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleTextResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'text/plain; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: text.toString().length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(text.toString());
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -773,56 +674,39 @@ describe('handleTextResponse()', () => {
 
 describe('handleHtmlResponse()', () => {
   describe('with a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const html = '<h1>abc</h1>';
       const response = {
         status: HTTP_STATUS_OK,
         html,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleHtmlResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_OK,
-        [HTTP2_HEADER_CONTENT_TYPE]: 'text/html; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: html.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(html);
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('without a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const html = '<h1>abc</h1>';
       const response = {
         html,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleHtmlResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'text/html; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: html.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(html);
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -833,82 +717,57 @@ describe('handleHtmlResponse()', () => {
 
 describe('handleJsonResponse()', () => {
   describe('with a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const json = { abc: 123 };
       const response = {
         status: HTTP_STATUS_OK,
         json,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleJsonResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_OK,
-        [HTTP2_HEADER_CONTENT_TYPE]: 'application/json; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: JSON.stringify(json).length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(JSON.stringify(json));
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('without a status property', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const json = { abc: 123 };
       const response = {
         json,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleJsonResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'application/json; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: JSON.stringify(json).length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(JSON.stringify(json));
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 
   describe('when passed a value that is already serialized', () => {
-    test('stream.respond() and stream.end() are called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const json = JSON.stringify({ abc: 123 });
       const response = {
         json,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
       const result = handleJsonResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'application/json; charset=UTF-8',
-        [HTTP2_HEADER_CONTENT_LENGTH]: json.length,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders);
-      expect(mockStream.end).toHaveBeenCalledWith(json);
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
@@ -919,103 +778,77 @@ describe('handleJsonResponse()', () => {
 
 describe('handleFileResponse()', () => {
   describe('with a status property', () => {
-    test('stream.respondWithFile() is called', () => {
+    test('response.end() is not called', async () => {
       const response = {
         status: HTTP_STATUS_OK,
         file: __filename,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleFileResponse(
+      const result = await handleFileResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
-      expect(outgoingHeaders).toHaveProperty(HTTP2_HEADER_STATUS, HTTP_STATUS_OK);
-      expect(mockStream.respondWithFile).toHaveBeenCalled();
+      expect(mockResponse.end).not.toHaveBeenCalled();
     });
   });
 
   describe('without a status property', () => {
-    test('stream.respondWithFile() is called', () => {
+    test('response.end() is not called', async () => {
       const response = {
         file: __filename,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
 
-      const result = handleFileResponse(
+      const result = await handleFileResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
-      expect(outgoingHeaders).not.toHaveProperty(HTTP2_HEADER_STATUS);
-      expect(mockStream.respondWithFile).toHaveBeenCalled();
+      expect(mockResponse.end).not.toHaveBeenCalled();
     });
   });
-});
 
-
-/**
- * statCheck()
- */
-
-describe('statCheck()', () => {
-  describe('adds the expected headers to the headers object', () => {
-    test('when passed an empty headers object', () => {
-      const headers = {};
-
-      //@ts-expect-error
-      const result = statCheck(mockStat, headers);
-
-      expect(result).toBeUndefined();
-      expect(headers).toHaveProperty(HTTP2_HEADER_CONTENT_TYPE, 'application/octet-stream');
-      expect(headers).toHaveProperty(HTTP2_HEADER_ACCEPT_RANGES, 'none');
-      expect(headers).toHaveProperty(HTTP2_HEADER_LAST_MODIFIED, mockDate.toUTCString());
-      expect(mockStat.mtime.toUTCString).toHaveBeenCalled();
-    });
-
-    test('when passed a headers object with properties already set', () => {
-      const dateStr = new Date().toUTCString();
-      const headers = {
-        [HTTP2_HEADER_CONTENT_TYPE]: 'text/plain',
-        [HTTP2_HEADER_ACCEPT_RANGES]: 'bytes',
-        [HTTP2_HEADER_LAST_MODIFIED]: dateStr,
+  describe('with `content-type`, `accept-ranges`, and `last-modified` headers', () => {
+    test('response.end() is not called', async () => {
+      const response = {
+        file: __filename,
       };
 
-      //@ts-expect-error
-      const result = statCheck(mockStat, headers);
+      const mockResponse = createMockResponse();
 
-      expect(result).toBeUndefined();
-      expect(headers).toHaveProperty(HTTP2_HEADER_CONTENT_TYPE, 'text/plain');
-      expect(headers).toHaveProperty(HTTP2_HEADER_ACCEPT_RANGES, 'bytes');
-      expect(headers).toHaveProperty(HTTP2_HEADER_LAST_MODIFIED, dateStr);
-      expect(mockStat.mtime.toUTCString).toHaveBeenCalled();
-    });
-  });
-});
+      mockResponse.setHeader(HTTP_HEADER_CONTENT_TYPE, 'application/octet-stream');
+      mockResponse.setHeader(HTTP_HEADER_ACCEPT_RANGES, 'none');
+      mockResponse.setHeader(HTTP_HEADER_LAST_MODIFIED, (new Date()).toUTCString());
 
-/**
- * onError()
- */
-
-describe('onError()', () => {
-  describe('when passed a stream and an error', () => {
-    test('stream.emit() is called with the expected argument', () => {
-      const err = Error('mock error') as NodeJS.ErrnoException;
-      const result = onError(
-        //@ts-expect-error
-        mockStream,
-        err,
+      const result = await handleFileResponse(
+        response,
+        mockResponse,
       );
 
       expect(result).toBeUndefined();
-      expect(mockStream.emit).toHaveBeenCalledWith('error', err);
+      expect(mockResponse.end).not.toHaveBeenCalled();
+    });
+  });
+
+  describe('with a length property', () => {
+    test('response.end() is not called', async () => {
+      const response = {
+        file: __filename,
+        length: 1
+      };
+      const mockResponse = createMockResponse();
+
+      const result = await handleFileResponse(
+        response,
+        mockResponse,
+      );
+
+      expect(result).toBeUndefined();
+      expect(mockResponse.end).not.toHaveBeenCalled();
     });
   });
 });
@@ -1026,24 +859,19 @@ describe('onError()', () => {
 
 describe('handleStatusResponse()', () => {
   describe('when passed a `status` argument of 418', () => {
-    test('stream.respond() is called with the expected arguments', () => {
+    test('response.end() is called', () => {
       const response = {
         status: HTTP_STATUS_TEAPOT,
       };
-      const outgoingHeaders = {};
+      const mockResponse = createMockResponse();
+
       const result = handleStatusResponse(
         response,
-        //@ts-expect-error
-        mockStream,
-        outgoingHeaders,
+        mockResponse,
       );
 
-      const expectedHeaders = {
-        [HTTP2_HEADER_STATUS]: HTTP_STATUS_TEAPOT,
-      };
-
       expect(result).toBeUndefined();
-      expect(mockStream.respond).toHaveBeenCalledWith(expectedHeaders, { endStream: true });
+      expect(mockResponse.end).toHaveBeenCalled();
     });
   });
 });
